requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
PyPDF2==3.0.1
google-generativeai==0.7.2
